Edit Test Settings fungerar, tidsnedräkning fungerar, småfix

+++ TestPlatform/src/TestPlatform/ViewModels/ViewQuestionVM.cs
@@ -11,7 +11,7 @@ namespace TestPlatform.ViewModels
         public string TestTitle { get; set; }
         public int QuestionIndex { get; set; }
         public int NumOfQuestion { get; set; }
-        public int? SecondsLeft { get; set; }
+        public double? SecondsLeft { get; set; }
         public QuestionFormVM QuestionFormVM { get; set; }
     }
 }
